Refactored "utils" library, reworked WRD command previewer
Did a decent bit of research into the opcodes, but it looks like "I've still got a ways to go".
